pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "writer-bot"
version = "0.1.0"
description = "Automated chapter writer CLI following the Universal AI Book-Writing Protocol."
authors = [{ name = "DevBot", email = "dev@example.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.14.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer[all]>=0.12",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "coverage",
    "build",
]

[project.scripts]
writer-bot = "writer_bot.__main__:app"

[tool.mypy]
strict = true
python_version = "3.11"

[tool.ruff]
line-length = 88
